Task 7 Finale I Hope

+++ 3/imgops.c
@@ -287,17 +287,17 @@ void normalize( uint8_t array[],
 {
   uint8_t oldmin = min(array, cols, rows);
   uint8_t oldmax = max(array, cols, rows);
-  uint8_t difference = oldmax-oldmin;
+  double difference = oldmax-oldmin;
   printf("min is: %d\n", oldmin);
   printf("max is: %d\n", oldmax);
-  printf("oldrange is: %d\n", difference);
+  printf("oldrange is: %f\n", difference);
 
   uint8_t newmax = 255;
   uint8_t newmin = 0;
 
   int i;
   int length = rows * cols;
-  uint8_t test[length];
+  double test[length];
   for (i=0; i<length; i++){
     test[i] = 0;
   }
